requirements.txt
Django==2.1
django-cors-headers==2.4.0
django-rest-swagger==2.2.0
django-summernote==0.8.10.0
djangorestframework==3.9.0
djangorestframework-jsonapi==2.6.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.3
djoser==1.3.2